+++ azure/datalake/store/__init__.py
@@ -6,7 +6,7 @@
 # license information.
 # --------------------------------------------------------------------------
 
-__version__ = "0.0.4"
+__version__ = "0.0.5"
 
 from .core import AzureDLFileSystem
 from .multithread import ADLDownloader